BMS: various refactors / remove some bloat
