+++ APP/BMS/BMS.c
@@ -6,63 +6,10 @@
  *  including ADC setup for current and voltage sensing, GPIO configuration for protection flags,
  *  PWM control for battery heaters, and I2C communications for telemetry and telecommand.
  *
- *  ADC channels are assigned to specific pins and memory buffers for multi-channel sweeps.
- *  GPIO pins are configured for over-voltage, under-voltage, and over-current protection flags.
- *  PWM outputs control battery heaters, with flexible duty cycle settings.
- *  Real-Time Clock (RTC) is used for periodic wakeup and routine execution.
- *  Telemetry and telecommand are handled via a custom protocol (tinyprotocol) over I2C.
- *
- *  * --- BMS.c Code Walkthrough ---
- *
- * This file implements the Battery Management System (BMS) logic for the MSP430.
- * 
- * Key sections:
- * 
- * 1. Hardware Definitions:
- *    - ADC pin assignments: Each *_PIN macro maps a physical pin to an ADC input channel.
- *    - ADC memory buffers: Each *_MEM macro maps an ADC channel to a memory buffer. These buffers store the results of ADC conversions.
- *    - GPIO pin assignments: Unique macros for each protection flag (over-voltage, under-voltage, over-current) mapped to specific pins.
- *    - PWM heater pins: Defines which pins are used for heater PWM outputs.
- *    - External ADC and I2C addresses.
- *
- * 2. Static Data Buffers:
- *    - Buffers for system status, current, voltage, flags, and external ADC readings.
- *    - These are used for telemetry responses and protocol communication.
- *
- * 3. Initialization Functions:
- *    - initADCs(): Sets up all ADC channels and memory buffers for multi-channel sweeps.
- *    - initGPIO(): Configures all GPIO pins for flags, I2C, and PWM outputs.
- *    - initClockTo16MHz(): Sets the MSP430 clock to 16MHz for fast operation.
- *    - initRTC(): Initializes the Real-Time Clock for periodic interrupts.
- *    - initBSP(): Calls all hardware initialization routines.
- *
- * 4. Communication Setup:
- *    - InitAppComm(): Initializes I2C and the custom tinyprotocol for telemetry/telecommand.
- *    - Registers all telemetry channels and telecommands.
- *
- * 5. Main Periodic Routine:
- *    - RoutineCycle_Process(): Called by RTC interrupt. Sweeps ADCs, reads GPIO flags, collects external ADC data, and updates buffers.
- *
- * 6. Interrupt Service Routine:
- *    - RTC_ISR(): Handles RTC interrupts, triggers periodic routine, and blinks LED for events.
- *
- * 7. Telemetry and Telecommand Handlers:
- *    - SendTelemetryResponse(): Responds to telemetry requests.
- *    - ProcessTelemetryRequest(): Dispatches telemetry responses.
- *    - SendPWM(): Sets heater PWM duty cycles.
- *    - ProcessTelecommand(): Handles incoming telecommands (e.g., heater control).
- *
- * --- Notes for Reviewers ---
- * - ADC memory buffers are crucial for storing conversion results; each sensor channel has a dedicated buffer.
- * - GPIO flag assignments should be unique and clearly mapped to physical pins.
- * - All hardware initialization is grouped for clarity and maintainability.
- * - Telemetry/telecommand protocol is modular and easily extendable.
- * - RoutineCycle_Process is the main data acquisition and update loop, triggered by RTC.
- * - Comments throughout the file explain hardware mapping and logic flow.
  *  Author: Brendan Kelly
  */
 
-#include <stdint.h>
+#include <stdbool.h>
 
 #include "BMS.h"
 #include "ADC_Read.h"
@@ -72,6 +19,10 @@
 #include "PWM.h"
 #include "swi2c.h"
 #include "ads7138irter.h"
+#include "tinyprotocol.h"
+#include "i2c.h"
+#include "bms_types.h"
+
 #if defined (__MSP430FR5989__)
 #include "rtc_c.h"
 #elif defined (__MSP430FR5969__)
@@ -308,86 +259,30 @@ static void initGPIO()
     PWM_PinSelect(HEATER_PWM_PORT, HEATER4_PWM_PIN);
 }
 
-
-
 // Initialize the Real-Time Clock (RTC) for periodic interrupts
 // Uses RTC_C or RTC_B driverlib depending on device
-#if defined (__MSP430FR5989__)
 static void initRTC()
 {
-    Calendar currentTime;
-
-    //Setup for Calendar
-    currentTime.Seconds    = 0x00;
-    currentTime.Minutes    = 0x00;
-    currentTime.Hours      = 0x00;
-    currentTime.DayOfWeek  = 0x00;
-    currentTime.DayOfMonth = 0x00;
-    currentTime.Month      = 0x00;
-    currentTime.Year       = 0x7E9;  // 2025
-
-    //Initialize Calendar Mode of RTC
-    RTC_C_initCalendar(RTC_C_BASE, &currentTime, RTC_C_FORMAT_BCD);
-
-    //Setup Calendar Alarm for 30 minutes after start.
-    RTC_C_configureCalendarAlarmParam param = {0};
-    param.minutesAlarm      = 0x2;  // Currently set to 2 minute for testing - TODO change to 24 hours
-    param.hoursAlarm        = 0x0;
-    param.dayOfWeekAlarm    = 0x0;
-    param.dayOfMonthAlarm   = 0x0;
-    RTC_C_configureCalendarAlarm(RTC_C_BASE, &param);
-
+#if defined (__MSP430FR5989__)
     RTC_C_clearInterrupt(RTC_C_BASE,
         RTC_C_CLOCK_READ_READY_INTERRUPT +
         RTC_C_TIME_EVENT_INTERRUPT +
         RTC_C_CLOCK_ALARM_INTERRUPT
         );
-    //Enable interrupt for RTC Ready Status, which asserts when the RTC
-    //Calendar registers are ready to read.
-    //Also, enable interrupts for the Calendar alarm and Calendar event.
     RTC_C_enableInterrupt(RTC_C_BASE,
         RTC_C_CLOCK_READ_READY_INTERRUPT +
         RTC_C_TIME_EVENT_INTERRUPT +
         RTC_C_CLOCK_ALARM_INTERRUPT
     );
 
-    //Start RTC Clock
+    //Start RTC
     RTC_C_startClock(RTC_C_BASE);
-}
-
 #elif defined (__MSP430FR5969__)
-static void initRTC()
-{
-    Calendar currentTime;
-
-    //Setup for Calendar
-    currentTime.Seconds    = 0x00;
-    currentTime.Minutes    = 0x00;
-    currentTime.Hours      = 0x00;
-    currentTime.DayOfWeek  = 0x00;
-    currentTime.DayOfMonth = 0x00;
-    currentTime.Month      = 0x00;
-    currentTime.Year       = 0x7E9;  // 2025
-
-    //Initialize Calendar Mode of RTC
-    RTC_B_initCalendar(RTC_B_BASE, &currentTime, RTC_B_FORMAT_BCD);
-
-    //Setup Calendar Alarm for 30 minutes after start.
-    RTC_B_configureCalendarAlarmParam param = {0};
-    param.minutesAlarm      = 0x2;  // Currently set to 2 minute for testing - TODO change to 24 hours
-    param.hoursAlarm        = 0x0;
-    param.dayOfWeekAlarm    = 0x0;
-    param.dayOfMonthAlarm   = 0x0;
-    RTC_B_configureCalendarAlarm(RTC_B_BASE, &param);
-
     RTC_B_clearInterrupt(RTC_B_BASE,
         RTC_B_CLOCK_READ_READY_INTERRUPT +
         RTC_B_TIME_EVENT_INTERRUPT +
         RTC_B_CLOCK_ALARM_INTERRUPT
         );
-    //Enable interrupt for RTC Ready Status, which asserts when the RTC
-    //Calendar registers are ready to read.
-    //Also, enable interrupts for the Calendar alarm and Calendar event.
     RTC_B_enableInterrupt(RTC_B_BASE,
         RTC_B_CLOCK_READ_READY_INTERRUPT +
         RTC_B_TIME_EVENT_INTERRUPT +
@@ -396,9 +291,8 @@ static void initRTC()
 
     //Start RTC Clock
     RTC_B_startClock(RTC_B_BASE);
-}
 #endif
-
+}
 
 // Private helper functions for telemetry and telecommand processing
 static void I2C_Proc_RX_Data(uint8_t data);
@@ -534,7 +428,7 @@ void BMS_init()
  * NOTE: inlining functions is entirely up to the compiler, therefore logic for different buffers (i.e. ADC, GPIO, etc) is kept in one function
  *  due to paranoia that the compiler might not inline it. This is subject to refactor. It is possible to force inline or use macros.
  */  
-extern void BMS_collectData()
+void BMS_collectData()
 {
     /* MSP430xxxx ON DEVICE ADCs */
     // Collect ADC data and put it into buffers
@@ -637,7 +531,7 @@ void RTC_ISR (void)
         case RTCIV_RTCRDYIFG: 
             sISRTriggered = true; 
             break;
-        case RTCIV_RTCTEVIFG:    P1OUT |= BIT0; break; // I think this is blinking an LED or some shit
+        case RTCIV_RTCTEVIFG:    break; 
         case RTCIV_RTCAIFG:      /* alarm */ break;
         case RTCIV_RT0PSIFG:     /* prescale 0 */ break;
         case RTCIV_RT1PSIFG:     /* prescale 1 */ break;
