+++ APP/BMS/BMS.h
@@ -2,171 +2,38 @@
  * @file
  * @brief Battery Management System (BMS) application interface for MSP430.
  *
- * @details
- * This module provides hardware abstraction and telemetry/telecommand definitions for
- * battery monitoring and control. It includes ADC channel assignments, protection flag
- * mapping, and protocol response structures for system status, current, voltage, and
- * heater control. Initialization routines set up all hardware and register protocol
- * channels for communication with the central data handler.
- *
  * @author Brendan Kelly
  */
 
 #ifndef BMS_H
 #define BMS_H
 
-#include <stdint.h>
 #include <stdbool.h>
 
-#include "tinyprotocol.h"
-#include "i2c.h"
-
-
-/**
- * @def SYSTEM_STATUS_RESP_LEN
- * @brief Number of bytes in system status telemetry response.
- */
-#define SYSTEM_STATUS_RESP_LEN   5
-
-/**
- * @def CURRENT_VOLTAGE_RESP_LEN
- * @brief Number of bytes in current or voltage telemetry response.
- */
-#define CURRENT_VOLTAGE_RESP_LEN 4
-
-/**
- * @def ADC_NUM_CHANNELS
- * @brief Number of ADC channels used for battery monitoring.
- */
-#define ADC_NUM_CHANNELS         8
-
-/**
- * @enum BMS_TelemetryRequestCmdId
- * @brief Command IDs for BMS telemetry requests.
- * @note If you don't manually assign these values, bad things happen that I don't understand.
- */
-enum BMS_TelemetryRequestCmdId {
-  BMS_SYSTEM_STATUS_ID        = 1,  /**< System status telemetry */
-  BMS_HEALTH_CHECK_ID         = 2,  /**< Health check telemetry */
-  BMS_FLAG_ID                 = 3,  /**< Protection flags telemetry */
-  BMS_CURRENT_DRAW_ID         = 4,  /**< Discharge current telemetry */
-  BMS_CURRENT_CHARGE_ID       = 5,  /**< Charge current telemetry */
-  BMS_VOLTAGE_BATTERY1_ID     = 6,  /**< Battery 1 voltage telemetry */
-  BMS_VOLTAGE_BATTERY2_ID     = 7,  /**< Battery 2 voltage telemetry */
-  BMS_VOLTAGE_COMBINED_ID     = 8,  /**< Combined battery voltage telemetry */
-  BMS_HEATERS_CONTROLLER_ID   = 9,  /**< Heater controller telecommand */
-  BMS_THERMISTOR03_DATA_ID    = 10, /**< Thermistor 0-3 external ADC telemetry */
-  BMS_THERMISTOR47_DATA_ID    = 11, /**< Thermistor 4-7 external ADC telemetry */
-};
-
-/**
- * @union SystemStatusResp_t
- * @brief Telemetry response for system status.
- */
-typedef union SystemStatusResp {
-    struct {
-        uint32_t runtime;    /**< System runtime in seconds */
-        uint8_t fw_version;  /**< Firmware version */
-    };
-    uint8_t buffer[SYSTEM_STATUS_RESP_LEN]; /**< Raw buffer for protocol transmission */
-} SystemStatusResp_t;
-
-/**
- * @union CurrentResp_t
- * @brief Telemetry response for current sensing.
- */
-typedef union CurrentResp {
-    struct {
-        uint16_t isense1; /**< Current sensor 1 reading */
-        uint16_t isense2; /**< Current sensor 2 reading */
-    };
-    uint8_t buffer[CURRENT_VOLTAGE_RESP_LEN]; /**< Raw buffer for protocol transmission */
-} CurrentResp_t;
-
-/**
- * @union VoltageResp_t
- * @brief Telemetry response for voltage sensing.
- */
-typedef union VoltageResp {
-    struct {
-        uint16_t vcell_a; /**< Voltage cell A reading */
-        uint16_t vcell_b; /**< Voltage cell B reading */
-    };
-    uint8_t buffer[CURRENT_VOLTAGE_RESP_LEN]; /**< Raw buffer for protocol transmission */
-} VoltageResp_t;
-
-/**
- * @union CombinedVoltageResp_t
- * @brief Telemetry response for combined battery voltages.
- */
-typedef union CombinedVoltageResp {
-    struct {
-        uint16_t vbatt1; /**< Battery 1 voltage */
-        uint16_t vbatt2; /**< Battery 2 voltage */
-    };
-    uint8_t buffer[CURRENT_VOLTAGE_RESP_LEN]; /**< Raw buffer for protocol transmission */
-} CombinedVoltageResp_t;
-
-/**
- * @union Flag_t
- * @brief Telemetry response for protection flags.
- */
-typedef union Flag {
-    uint16_t val;      /**< Packed flag value (bitfield) */
-    uint8_t buffer[2]; /**< Raw buffer for protocol transmission */
-} Flag_t;
+extern const struct TINYPROTOCOL_Config protocolConfig;
 
 /**
  * @brief Initializes the Battery Management System hardware and communication.
  *
  * Sets up ADCs, GPIOs, PWM, RTC, and registers telemetry/telecommand channels.
- *  
- * @code
- * int main(void)
- * {
- *     BMS_init();
- * 
- *     __bis_SR_register(GIE);
- *     while (1) {
- *         if (BMS_ISRTriggered()) {
- *             BMS_collectData();
- *         }
- *     }
- *     return 0;
- * }
- * @endcode
- */
-extern void BMS_init();
 
+ */
+void BMS_init();
 
 /**
  * @brief Synchronous data-acquisition pass that refreshes telemetry buffers.
  *
- * Starts one ADC12_B sequence-of-channels sweep (MEM0→MEM9) and *busy-waits*
- * until it completes, then copies results into current/voltage unions. Also
- * samples protection-flag GPIOs and reads 8 channels from the external
- * ADS7138 (thermistors) over software I²C into the two 8-byte buffers.
- *
- * @details
- * - **Synchronous/blocking:** contains a busy-wait on ADC completion.
- * - **Side effects:** clears the internal “RTC fired” flag exposed by
- *   ::BMS_ISRTriggered(), acknowledging the event, and modifies internal static 
- *   data buffers to be read by master (CDH) via Tiny Protocol
- *
  * @pre ::BMS_init() has been called.
  * @post Telemetry buffers registered with tinyprotocol contain fresh data.
  */
-extern void BMS_collectData();
+void BMS_collectData();
 
 /**
  * @brief Check whether an RTC event occurred since the last collection.
  *
- * This returns a latch set by the RTC ISR. The latch is cleared by
- * ::BMS_collectData() after it completes a refresh cycle.
- *
  * @return `true` if an RTC event has occurred and data collection is due;
  *         `false` otherwise.
  */
-extern bool BMS_ISRTriggered();
+bool BMS_ISRTriggered();
 
 #endif
